move `formatScores` to `lib/leaderboard/scores.ts`
